# analysis-service/app/main.py
"""
OpenSight Analiz Servisi (FastAPI) - RabbitMQ'den gelen trafik olaylarını tüketiyor,
Redis'te rolling pencere tutuyor, RollingZScoreDetector + BehavioralAnomalyDetector
ile paralel skorluyor, CorrelationEngine'de birleştiriyor ve ExplanationGenerator ile
açıklama üretip OpenSight.Api'ye yazıyor
"""
from __future__ import annotations

import logging
import os
import threading
from dataclasses import asdict

# basicConfig olmadan logger.info(...) hiçbir yere yazılmıyordu - Render'da RabbitMQ bağlantı logu yokluğu buydu
logging.basicConfig(level=logging.INFO, format="%(asctime)s %(levelname)s %(name)s: %(message)s")

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field
from redis import Redis

from app.messaging.backend_client import BackendClient
from app.messaging.rabbitmq_consumer import RabbitMqTrafficConsumer
from app.services.anomaly_pipeline import AnomalyPipeline
from app.services.behavioral_detector import BehavioralAnomalyDetector
from app.services.cold_start import ColdStartManager
from app.services.correlation_engine import CorrelationEngine
from app.services.explanation_generator import ExplanationGenerator
from app.services.latency_metrics import LatencyMetrics
from app.services.performance_detector import RollingZScoreDetector
from app.services.threshold_settings import ThresholdSettingsService
from app.services.traffic_window import ClientTrafficWindow

REDIS_URL = os.environ.get("REDIS_URL", "")
REDIS_HOST = os.environ.get("REDIS_HOST", "localhost")
REDIS_PORT = int(os.environ.get("REDIS_PORT", "6379"))
RABBITMQ_URI = os.environ.get("RABBITMQ_URI", "")
RABBITMQ_HOST = os.environ.get("RABBITMQ_HOST", "localhost")
RABBITMQ_PORT = int(os.environ.get("RABBITMQ_PORT", "5672"))
# baseline yoğunun patlama dönemini de görmeli, o yüzden pencere patlama aralığından (~120-240s) uzun
COLD_START_SECONDS = int(os.environ.get("COLD_START_SECONDS", "420"))
# aynı istemci + aynı tür için bu süre dolmadan yeni alarm yazılmıyor
ALERT_COOLDOWN_SECONDS = float(os.environ.get("ALERT_COOLDOWN_SECONDS", "300"))
BACKEND_URL = os.environ.get("OPENSIGHT_API_URL", "http://localhost:8080")
# Render'da CORS_ORIGIN olarak girildi, yerelde/docker-compose'da hala DASHBOARD_ORIGIN kullaniliyor
DASHBOARD_ORIGIN = os.environ.get("CORS_ORIGIN") or os.environ.get("DASHBOARD_ORIGIN", "http://localhost:5173")
# boşsa Ollama kapalı, sadece şablon açıklama kullanılıyor
OLLAMA_URL = os.environ.get("OLLAMA_URL", "")
OLLAMA_MODEL = os.environ.get("OLLAMA_MODEL", "llama3.2:1b")

app = FastAPI(title="OpenSight Analiz Servisi", version="0.1.0")


def _env_flag(name: str) -> bool:
    """her istekte okunuyor, canlıda env değişince yeniden başlatma dışında kod değişmesin"""
    return os.environ.get(name, "").strip().lower() in ("1", "true", "yes", "on")

# dashboard'un dogrudan tarayicidan (fetch ile) esik ayarlari endpoint'lerine erisebilmesi icin -
# backend'deki Cors:DashboardOrigin yaklasimiyla ayni mantik
app.add_middleware(
    CORSMiddleware,
    allow_origins=[DASHBOARD_ORIGIN],
    allow_methods= ["*"],
    allow_headers= ["*"],
)

# REDIS_URL verilmisse (Upstash gibi rediss:// TLS URL'i) from_url TLS'i de otomatik parse ediyor,
# yoksa eski host/port fallback'i (yerel gelistirme / docker-compose) kullaniliyor
if REDIS_URL:
    redis_client = Redis.from_url(REDIS_URL, decode_responses=True)
else:
    redis_client = Redis(host=REDIS_HOST, port=REDIS_PORT, decode_responses=True)

performance_detector = RollingZScoreDetector(redis_client, window_size=50, threshold=3.2)
# 0.05'te normal/yoğunun yarısı yanlış işaretleniyordu, 0.005'te şüpheli yine güvenilir ayrışıyor -
# dashboard'daki ThresholdConfig.tsx DEFAULT_CONTAMINATION ile aynı tutulmalı
DEFAULT_CONTAMINATION = 0.005
behavioral_detector = BehavioralAnomalyDetector(contamination=DEFAULT_CONTAMINATION)
traffic_window = ClientTrafficWindow(redis_client)
cold_start = ColdStartManager(redis_client, behavioral_detector, cold_start_seconds=COLD_START_SECONDS)
correlation_engine = CorrelationEngine(window_seconds=30 * 60)
explanation_generator = ExplanationGenerator(ollama_url=OLLAMA_URL, model=OLLAMA_MODEL)
backend_client = BackendClient(BACKEND_URL)
latency_metrics = LatencyMetrics(redis_client)
threshold_settings = ThresholdSettingsService(
    redis_client, performance_detector, behavioral_detector, baseline_reader=cold_start.read_baseline
)

anomaly_pipeline = AnomalyPipeline(
    performance_detector, behavioral_detector, traffic_window, cold_start, correlation_engine, backend_client,
    threshold_settings=threshold_settings,
    explanation_generator=explanation_generator,
    latency_metrics=latency_metrics,
    alert_cooldown_seconds=ALERT_COOLDOWN_SECONDS,
)
traffic_consumer = RabbitMqTrafficConsumer(RABBITMQ_HOST, anomaly_pipeline, port=RABBITMQ_PORT, uri=RABBITMQ_URI or None)
_consumer_thread: threading.Thread | None = None


@app.on_event("startup")
def start_traffic_consumer() -> None:
    """backend'i ısıtıyor, sonra RabbitMQ consumer'ı arka plan thread'inde başlatıyor - API'yi bloklamıyor"""
    backend_client.warmup()
    global _consumer_thread
    _consumer_thread = threading.Thread(target=traffic_consumer.run_forever, daemon=True)
    _consumer_thread.start()


@app.on_event("shutdown")
def stop_traffic_consumer() -> None:
    traffic_consumer.stop()


@app.get("/health")
def health() -> dict:
    return {"status": "healthy", "service": "opensight-analysis"}


@app.get("/status")
def status() -> dict:
    """modelin eğitilip eğitilmediğini döndürüyor - simülatör şüpheli trafiği bu sinyalle başlatıyor"""
    return {
        "model_ready": cold_start.is_ready(),
        "in_cold_start": cold_start.is_in_cold_start(),
        "baseline_samples": cold_start.baseline_size(),
        "baseline_scale": cold_start.baseline_scale(),
        "last_trained_at": behavioral_detector.last_trained_at,
    }


@app.get("/metrics/latency")
def get_latency_metrics(minutes: int = 30, client_id: str | None = None) -> dict:
    """son N dakikanın dakikalık ortalama gecikmesi ve istek sayısı, client_id verilirse sadece o istemci"""
    return latency_metrics.series(minutes=minutes, client_id=client_id)


class LatencySample(BaseModel):
    client_id: str
    latency_ms: float


@app.post("/debug/score-latency", include_in_schema=False)
def score_latency(sample: LatencySample) -> dict:
    """gerçek RabbitMQ akışı olmadan rolling z-score'u test etmek için, ENABLE_DEBUG_ENDPOINTS=true değilse 404"""
    if not _env_flag("ENABLE_DEBUG_ENDPOINTS"):
        raise HTTPException(status_code=404, detail="Not Found")
    result = performance_detector.update_and_score(sample.client_id, sample.latency_ms)
    return {
        "z_score": result.z_score,
        "is_anomaly": result.is_anomaly,
        "mean": result.mean,
        "std": result.std,
    }


class ThresholdSettingsResponse(BaseModel):
    z_score_threshold: float
    contamination: float
    last_trained_at: str | None
    alert_counts_last_24h: dict[str, int]
    read_only: bool = False


class UpdateThresholdSettingsRequest(BaseModel):
    z_score_threshold: float | None =Field(default=None, gt=0)
    contamination: float | None = Field(default=None, gt=0, le=0.5)


@app.get("/settings/thresholds", response_model=ThresholdSettingsResponse)
def get_threshold_settings()-> ThresholdSettingsResponse:
    """Z-Score eşiğini, contamination oranını, son 24 saatteki alarm sayısını ve son eğitim zamanını döndürüyor"""
    return ThresholdSettingsResponse(
        **asdict(threshold_settings.get_settings()), read_only=_env_flag("SETTINGS_READ_ONLY")
    )


@app.put("/settings/thresholds", response_model=ThresholdSettingsResponse)
def update_threshold_settings(request: UpdateThresholdSettingsRequest) -> ThresholdSettingsResponse:
    """yeni eşikleri canlı detector'lara uyguluyor ve Redis'e kalıcı olarak yazıyor, SETTINGS_READ_ONLY=true ise 403"""
    if _env_flag("SETTINGS_READ_ONLY"):
        raise HTTPException(status_code=403, detail="Bu ortamda ayarlar salt okunur (SETTINGS_READ_ONLY), değiştirilemez.")
    settings = threshold_settings.update_settings(request.z_score_threshold, request.contamination)
    return ThresholdSettingsResponse(**asdict(settings), read_only=False)