# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: OpenSight API
  version: "0.1.0"
  description: >-
    Mock Open Banking endpoint'leri ile alert ve korelasyon yönetimi.
    /status ve /settings/* analiz servisinde, /demo/* simülatörde (path bazında server tanımlı).
servers:
  - url: http://localhost:8080
paths:
...
  /settings/thresholds:
    servers:
      - url: http://localhost:8001
        description: Analiz servisi
    get:
      summary: Z-Score eşiği, contamination oranı ve son 24 saatteki alarm sayılarını döndürür
      responses:
        "200":
          description: Eşik ayarları
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThresholdSettings' }
    put:
      summary: Eşikleri canlı detector'lara uygular ve Redis'e yazar
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateThresholdSettingsRequest' }
      responses:
        "200":
          description: Güncel eşik ayarları
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThresholdSettings' }
        "403": { description: Ortamda ayarlar salt okunur (SETTINGS_READ_ONLY) }
        "422": { description: Geçersiz değer }
...
components:
...
  schemas:
...
    ThresholdSettings:
      type: object
      properties:
        z_score_threshold: { type: number }
        contamination: { type: number }
        last_trained_at: { type: string, format: date-time, nullable: true }
        alert_counts_last_24h:
          type: object
          additionalProperties: { type: integer }
        read_only: { type: boolean, description: "true ise PUT 403 döner" }
    UpdateThresholdSettingsRequest:
      type: object
      properties:
        z_score_threshold: { type: number, exclusiveMinimum: true, minimum: 0, nullable: true }
        contamination: { type: number, exclusiveMinimum: true, minimum: 0, maximum: 0.5, nullable: true }
...